(Bugfix)[Updated migarations and fixed crashing issues of the docker container while building]
- We are now tracking all the migration files for better consistency.
- updated docker compose to read existing migration files rather than trying to create a new one every time.
- Fixed issues with seed_demo.py which was causing the backend container to immediately crash without logs.
- Migration files are no longer gitignored

+++ backend/chat_app/management/commands/seed_demo.py
@@ -111,16 +111,16 @@ class Command(BaseCommand):
         # This could lead to dangling vector embeddings since they are linked to a separate DB using the user_id, and cascading delete won't work across DBs.
         plwd = self.get_or_create_demo_user("demo_patient",   password="1", first_name="John", last_name="Patient"  )
         care = self.get_or_create_demo_user("demo_caregiver", password="1", first_name="Jane", last_name="Caregiver", is_staff=True)
-        plwd_account = Account.objects.create(user=plwd, role="patient")
-        care_account = Account.objects.create(user=care, role="caregiver")
-        profile = Profile.objects.create(account=plwd_account, zipcode="9999", birthDate=timezone.now(), locationStatus="alone")
+        plwd_account, _ = Account.objects.get_or_create(user=plwd, defaults={"role": "patient"})
+        care_account, _ = Account.objects.get_or_create(user=care, defaults={"role": "caregiver"})
+        profile, _ = Profile.objects.get_or_create(account=plwd_account,defaults={"zipcode": "9999", "birthDate": timezone.now(), "locationStatus": "alone",},)
         
         # Create an Access object so caregiver account can access the plwd account
-        Access.objects.create(account=care_account, profile=profile)
+        Access.objects.get_or_create(account=care_account, profile=profile)
 
         # Also create settings and goal objects for the new Profile
-        UserSettings.objects.create(profile=profile)
-        Goal        .objects.create(profile=profile, target=5, start_date=two_days_ago)
+        UserSettings.objects.get_or_create(profile=profile)
+        Goal        .objects.get_or_create(profile=profile, target=5, start_date=two_days_ago)
         # Add sample ChatSessions
         self.seed_chats(profile, days_back=10)
         
@@ -132,15 +132,15 @@ class Command(BaseCommand):
         # --------------------------------------------------------------------
         plwd_2 = self.get_or_create_demo_user("buddy_user", password="1", first_name="Buddy", last_name="Robot"    )
         care_2 = self.get_or_create_demo_user("buddy_care", password="1", first_name="Buddy", last_name="Caregiver")
-        plwd_account_2 = Account.objects.create(user=plwd_2, role="patient")
-        care_account_2 = Account.objects.create(user=care_2, role="caregiver")
-        profile_2 = Profile.objects.create(account=plwd_account_2, zipcode="9999", birthDate=timezone.now(), locationStatus="alone")
+        plwd_account_2, _ = Account.objects.get_or_create(user=plwd_2, defaults={"role": "patient"})
+        care_account_2, _ = Account.objects.get_or_create(user=care_2, defaults={"role": "caregiver"})
+        profile_2, _ = Profile.objects.get_or_create(account=plwd_account_2, defaults={"zipcode": "9999", "birthDate": timezone.now(), "locationStatus": "alone",})
         
         # Create an Access object so caregiver account can access the plwd account
-        Access.objects.create(account=care_account_2, profile=profile_2)
+        Access.objects.get_or_create(account=care_account_2, profile=profile_2)
 
-        UserSettings.objects.create(profile=profile_2)
-        Goal        .objects.create(profile=profile_2, target=5, start_date=two_days_ago)
+        UserSettings.objects.get_or_create(profile=profile_2)
+        Goal        .objects.get_or_create(profile=profile_2, target=5, start_date=two_days_ago)
         self.seed_chats(profile_2, days_back=10)
         self.seed_reminders(profile_2, num_reminders=5)
         self.seed_activities()
@@ -150,7 +150,11 @@ class Command(BaseCommand):
     # Helper: Deletes all data associated with a user, such as Account, Profile, Settings, etc
     # ====================================================================
     def delete_user_data(self, user):
-        account = Account.objects.get(user=user)
+        account = Account.objects.filter(user=user).first()
+        # if there is no account linked to this user, then there is no data to delete, so we can just return.
+        if not account:
+            return
+
         RAGInstructions.objects.filter(user=user).delete()
         try: # The case where this user's account is the main account of a Profile object
             profile = Profile.objects.get(account=account)
@@ -310,10 +314,13 @@ class Command(BaseCommand):
         for idx, name in enumerate(DEMO_RAG_NAMES):
             description = DEMO_RAG_DESCRIPTIONS[idx]
             instructions = DEMO_RAG_INSTRUCTIONS[idx]
-            RAGInstructions.objects.create(
+            RAGInstructions.objects.update_or_create(
                 name=name,
-                description=description,
-                instructions=instructions,
                 user=demo_user,
                 activity=memory_activity,
+                defaults={
+                    "description": description,
+                    "instructions": instructions,
+                    "instruction_order": 1,
+                },
             )
